+++ SomnusFrontEnd/src/app/models/role.model.ts
@@ -1,5 +1,6 @@
 export enum Role {
     User = 'User',
     Editor = 'Editor',
-    Admin = 'Admin'
+    Admin = 'Admin',
+    Manager = 'Manager'
 }